Version 1.5
Fixed issue with different screen sizes.
Revamped the post view to help save on data!
Now, instead of loading the web view, the rss feed for that post is loaded, with a price, small photo, date, and description (if provided by the poster). If you really like the item, you can still click to open it up in a web browser and view the full page!